Statistics Javalin routes
